refactor(auth): rename handleNewUser.ts file to handle-new-user.ts for clarity
