+++ server/src/controllers/auth/handle-new-user.ts
