+++ server/src/controllers/auth/index.ts
@@ -1,1 +1,1 @@
-export * from './HandleNewUser';
+export * from "./handle-new-user";
